CoinbaseCharts: Fix fickering price metric values
Incorrect price metric values can be seen for a fraction of a second
when switching in between either the duration or cryptocurrency tabs.

This due to our component using the duration/cryptocurrency values
of our new state to compute the metrics when we are in the midst of
fetching the new set of price history data.

A new state variable 'activeSpotPrice' is added to store the spot
price used for computing the price metrics. This allow us to use
the old spot price to compute the metric values will we're fetching
the new set of price history data, effectively removing the
'flickering' effect.